+++ src/components/main-page-empty/main-page-empty.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import {FilteringType} from "../../const";
+import {FilteringType} from "../../consts/common";
 
 type Props = {
   currentCity: FilteringType;
