+++ src/components/map/map.tsx
@@ -1,7 +1,7 @@
 import Leaflet from "leaflet";
 import React from "react";
 
-import {MapIconUrl, MapIconSize, MapLayer, MapType} from "../../const";
+import {MapIconUrl, MapIconSize, MapLayer, MapType} from "../../consts/components";
 import {Offer} from "../../types";
 
 import "leaflet/dist/leaflet.css";
