+++ src/components/offers-list/offers-list.tsx
@@ -3,7 +3,7 @@ import {useDispatch} from "react-redux";
 
 import {getActiveOfferID} from "../../store/app/actions";
 
-import {CardType} from "../../const";
+import {CardType} from "../../consts/components";
 import {Offer} from "../../types";
 
 import OffersItem from "../offers-item/offers-item";
