+++ src/components/page-footer/page-footer.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {Link} from "react-router-dom";
 
-import {AppRoute} from "../../const";
+import {AppRoute} from "../../consts/common";
 
 const PageFooter: React.FunctionComponent = () => {
