+++ src/components/page-header/page-header.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import {Link} from "react-router-dom";
 
-import {AppRoute, HeaderType} from "../../const";
+import {AppRoute} from "../../consts/common";
+import {HeaderType} from "../../consts/components";
 
 import UserNav from "../user-nav/user-nav";
