+++ src/components/review-form/review-form.tsx
@@ -7,7 +7,7 @@ import {sendReview} from "../../store/data/api-actions";
 
 import {selectIsDataSending, selectIsSendingError} from "../../store/data/selectors";
 
-import {RATING_STARS_COUNT} from "../../const";
+import {RATING_STARS_COUNT} from "../../consts/components";
 import {getRatingArray, validateComment} from "../../utils/components";
 
 type Props = {
