+++ src/components/sorting-form/sorting-form.tsx
@@ -3,7 +3,7 @@ import {useDispatch} from "react-redux";
 
 import {getCurrentSorting} from "../../store/app/actions";
 
-import {SortingType} from "../../const";
+import {SortingType} from "../../consts/common";
 
 type Props = {
   currentSorting: SortingType;
