+++ src/middlewares/redirect.ts
@@ -1,7 +1,7 @@
 import {Middleware} from "redux";
 
 import history from "../history";
-import {ActionType} from "../store/const";
+import {ActionType} from "../consts/store";
 
 export const redirect: Middleware = () => (next) => (action) => {
   if (action.type === ActionType.REDIRECT_TO_ROUTE) {
