+++ src/store/app/actions.ts
@@ -1,6 +1,6 @@
 import {AppAction} from "./types";
-import {ActionType} from "../const";
-import {FilteringType, SortingType} from "../../const";
+import {FilteringType, SortingType} from "../../consts/common";
+import {ActionType} from "../../consts/store";
 
 export const getActiveOfferID = (activeOfferID: number | null): AppAction => ({
   type: ActionType.GET_ACTIVE_OFFER_ID,
