+++ src/store/app/reducer.ts
@@ -1,6 +1,6 @@
 import {AppAction, AppState} from "./types";
-import {ActionType} from "../const";
-import {FilteringType, SortingType} from "../../const";
+import {FilteringType, SortingType} from "../../consts/common";
+import {ActionType} from "../../consts/store";
 
 const initialState: AppState = {
   activeOfferID: null,
