+++ src/store/data/actions.ts
@@ -1,5 +1,5 @@
 import {DataAction} from "./types";
-import {ActionType} from "../const";
+import {ActionType} from "../../consts/store";
 import {Offer, Review} from "../../types";
 
 export const getAllOffers = (allOffers: Offer[]): DataAction => ({
