+++ src/store/data/reducer.ts
@@ -1,5 +1,5 @@
 import {DataAction, DataState} from "./types";
-import {ActionType} from "../const";
+import {ActionType} from "../../consts/store";
 import {Offer} from "../../types";
 import {getUpdatedFavoriteOffers, getUpdatedOffers} from "../../utils/store";
