+++ src/store/data/types.ts
@@ -1,4 +1,4 @@
-import {ActionType} from "../const";
+import {ActionType} from "../../consts/store";
 import {Offer, Review} from "../../types";
 
 export type DataState = {
