+++ src/store/user/actions.ts
@@ -1,6 +1,6 @@
 import {UserAction} from "./types";
-import {ActionType} from "../const";
-import {AppRoute, AuthStatus} from "../../const";
+import {AppRoute, AuthStatus} from "../../consts/common";
+import {ActionType} from "../../consts/store";
 import {User} from "../../types";
 
 export const getAuthStatus = (authStatus: AuthStatus): UserAction => ({
