+++ src/store/user/types.ts
@@ -1,5 +1,5 @@
-import {ActionType} from "../const";
-import {AppRoute, AuthStatus} from "../../const";
+import {AppRoute, AuthStatus} from "../../consts/common";
+import {ActionType} from "../../consts/store";
 import {User} from "../../types";
 
 export type UserState = {
