+++ src/utils/store.ts
@@ -1,4 +1,5 @@
-import {FavStatus, FilteringType, SortingType} from "../const";
+import {FilteringType, SortingType} from "../consts/common";
+import {FavStatus} from "../consts/store";
 import {Offer, Review} from "../types";
 
 export const getFavStatus = (isFavorite: boolean): string => `/${isFavorite ? FavStatus.NOT_FAV : FavStatus.FAV}`;
